delete auth added

+++ osa4/blogilista/controllers/blogs.js
@@ -3,22 +3,13 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const auth = request.get('authorization')
-  if (auth && auth.startsWith('Bearer ')) {
-    return auth.replace('Bearer ', '')
-  }
-  return null
-}
-
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  console.log("TOKEN", decodedToken)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'invalid token' })
   }
@@ -43,11 +34,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
   try {
-    const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
-    if (!deletedBlog) {
-      return response.status(404).json({ error: 'Blog not found' })
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'invalid token' })
     }
+    const deletedBlog = await Blog.findById(request.params.id)
+
+    if (decodedToken.id !== deletedBlog.user.toString())
+    {
+      return response.status(400).json({ error: 'not authorized' })
+    }
+    if (!deletedBlog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+    deletedBlog = await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   } catch (error) {
     response.status(400).json({ error: error.message })
